Setup searxng to run on AWS Lambda.
This builds a docker image based on the offical searxng image, but
packages it for lambda.

This can be built and run locally and brought up with docker-compose to
verify that things are working.

```sh
$ SEARXNG_SECRET_KEY=$(openssl rand -hex 32) docker compose up --build
```

And then you should have a searxng instance running at:

http://localhost:8002

I have not pinned all the docker images - I will let renovate do that
after committing.